Prettier script added and applied

+++ contracts/utils/IDateTime.sol
@@ -2,10 +2,15 @@
 pragma solidity ^0.8.0;
 
 interface IDateTime {
-  function getYear(uint timestamp) external pure returns (uint16);
-  function getMonth(uint timestamp) external pure returns (uint16);
-  function getDay(uint timestamp) external pure returns (uint16);
-  function getHour(uint timestamp) external pure returns (uint16);
-  function getMinute(uint timestamp) external pure returns (uint16);
-  function getSecond(uint timestamp) external pure returns (uint16);
+    function getYear(uint256 timestamp) external pure returns (uint16);
+
+    function getMonth(uint256 timestamp) external pure returns (uint16);
+
+    function getDay(uint256 timestamp) external pure returns (uint16);
+
+    function getHour(uint256 timestamp) external pure returns (uint16);
+
+    function getMinute(uint256 timestamp) external pure returns (uint16);
+
+    function getSecond(uint256 timestamp) external pure returns (uint16);
 }
